requirements.txt
RPi.GPIO
board
busio
adafruit-ssd1306
PIL
os
can
threading